create projet + view

+++ webapp/controller/CreateProject.controller.js
@@ -284,6 +284,7 @@ sap.ui.define([
 			this.getRouter().navTo("projects", {
 				ID: encodeURIComponent(oData.ID)
 			});
+			
 		},
 
 		/**
@@ -366,9 +367,9 @@ sap.ui.define([
 		 */		
 		_updateFUNC: function(){
 			var FUNC_field=this.byId("FUNC_id"),
-				FS=this._naNToZero(parseInt(this._removeSpaces(this.byId("FS_id").getProperty("value")))),
-				DEV_SUP=this._naNToZero(parseInt(this._removeSpaces(this.byId("DEV_SUP_id").getProperty("value")))),
-				FUT=this._naNToZero(parseInt(this._removeSpaces(this.byId("FUT_id").getProperty("value"))));
+				FS=this._naNToZero(parseFloat(this._removeSpaces(this.byId("FS_id").getProperty("value")))),
+				DEV_SUP=this._naNToZero(parseFloat(this._removeSpaces(this.byId("DEV_SUP_id").getProperty("value")))),
+				FUT=this._naNToZero(parseFloat(this._removeSpaces(this.byId("FUT_id").getProperty("value"))));
 			
 			var FUNC_value= FS+DEV_SUP+FUT;
 			
@@ -385,12 +386,12 @@ sap.ui.define([
 		 */	
 		_updateTECH: function(){
 			var TECH_field=this.byId("TECH_id"),
-				FS_SUP=this._naNToZero(parseInt(this._removeSpaces(this.byId("FS_SUP_id").getProperty("value")))),
-				FS_REV=this._naNToZero(parseInt(this._removeSpaces(this.byId("FS_REV_id").getProperty("value")))),
-				TS=this._naNToZero(parseInt(this._removeSpaces(this.byId("TS_id").getProperty("value")))),
-				DEV_UT=this._naNToZero(parseInt(this._removeSpaces(this.byId("DEV_UT_id").getProperty("value")))),
-				FUT_SUP=this._naNToZero(parseInt(this._removeSpaces(this.byId("FUT_SUP_id").getProperty("value")))),
-				FIT_SUP=this._naNToZero(parseInt(this._removeSpaces(this.byId("FIT_SUP_id").getProperty("value"))));
+				FS_SUP=this._naNToZero(parseFloat(this._removeSpaces(this.byId("FS_SUP_id").getProperty("value")))),
+				FS_REV=this._naNToZero(parseFloat(this._removeSpaces(this.byId("FS_REV_id").getProperty("value")))),
+				TS=this._naNToZero(parseFloat(this._removeSpaces(this.byId("TS_id").getProperty("value")))),
+				DEV_UT=this._naNToZero(parseFloat(this._removeSpaces(this.byId("DEV_UT_id").getProperty("value")))),
+				FUT_SUP=this._naNToZero(parseFloat(this._removeSpaces(this.byId("FUT_SUP_id").getProperty("value")))),
+				FIT_SUP=this._naNToZero(parseFloat(this._removeSpaces(this.byId("FIT_SUP_id").getProperty("value"))));
 			
 			var TECH_value=FS_SUP+FS_REV+TS+DEV_UT+FUT_SUP+FIT_SUP;
 			
@@ -403,10 +404,10 @@ sap.ui.define([
 		 * @private
 		 */	
 		_updateGrandTotal: function(){
-			var TECH=this._naNToZero(parseInt(this._removeSpaces(this.byId("TECH_id").getProperty("text")))),
-				FUNC=this._naNToZero(parseInt(this._removeSpaces(this.byId("FUNC_id").getProperty("text")))),
-				TECH_ARCH=this._naNToZero(parseInt(this._removeSpaces(this.byId("TECH_ARCH_id").getProperty("value")))),
-				TECH_LEAD=this._naNToZero(parseInt(this._removeSpaces(this.byId("TECH_LEAD_id").getProperty("value"))));
+			var TECH=this._naNToZero(parseFloat(this._removeSpaces(this.byId("TECH_id").getProperty("text")))),
+				FUNC=this._naNToZero(parseFloat(this._removeSpaces(this.byId("FUNC_id").getProperty("text")))),
+				TECH_ARCH=this._naNToZero(parseFloat(this._removeSpaces(this.byId("TECH_ARCH_id").getProperty("value")))),
+				TECH_LEAD=this._naNToZero(parseFloat(this._removeSpaces(this.byId("TECH_LEAD_id").getProperty("value"))));
 				
 			var GRAND_TOTAL=TECH+FUNC+TECH_ARCH+TECH_LEAD,
 				GRAND_TOTAL_field=this.byId("GRAND_TOTAL_id");
